src/requirements.txt
sqlmodel
alembic
asyncpg
uvicorn
fastapi
fastapi-health
python-decouple

# dev
pytest
pytest-cov
requests
pre-commit
